Register ContentObserver after the first data was delivered
If you register ContentObserver in the background thread before the data
is delivered, and the notifyChange on subscription Uri is called, the
data will be dropped and Loader will try to load the data again. It's a
good idea for further loads, because you do not want to show stale data,
but during first load it's desireable to show any data as soon as
possible.