Do not reload the data when loader is abandoned
When restartLoader is called, the new Loader instance is created in
onCreateLoader callback and the old instance is moved to abandoned
state. In this state Loader should keep the data until the newly created
Loader delivers the results, but it should not reload the data itself,
because the results would be dropped in onCancelled callback anyways.