Add failing test for a ComposedCursorLoader crash.
The exact nature of the bug was described in details in the test case.
While I *think* I can fix this stuff, I'm painfully aware that I have a
single test case for a single permutation of an order of asynchronous
operations that lead to the crash. The whole Loaders architecture is
nigh-untestable and requires in-depth knowledge of LoaderManagerImpl,
AsyncTaskLoader, and ModernAsyncTask) internals.

I'm investing the time to understand, test and fix this issue only
because it'll take a long time to replace 100% of Loaders in Base app
with something better (BTW: the best candidate right now is Rx, we might
consider also Kotlin coroutines, but only if it's possible to easily
test this kind of scenarios).